# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Мой гараж",
        "description": "Это список методов для интеграции с Мой гараж, в заголовках всех запросов должен присутствоваться API-ключ, который будет предоставлен администратором на почку клиента после регистрации клиента в сервисе администратором",
        "version": "1.0.2"
    },
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Получение данных пользователя (аутентифицированный запрос)",
                "description": "Получение данных пользователя (аутентифицированный запрос)",
                "operationId": "GetUser",
                "responses": {
                    "200": {
                        "description": "Успешная аутентификация или регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "description": "Данные пользователя",
                                            "properties": {
                                                "id": {
                                                    "description": "ID пользователя",
                                                    "type": "integer"
                                                },
                                                "code": {
                                                    "description": "Код пользователя",
                                                    "type": "integer"
                                                },
                                                "role_id": {
                                                    "description": "ID роли пользователя",
                                                    "type": "integer"
                                                },
                                                "user_status": {
                                                    "description": "Статус пользователя",
                                                    "type": "integer"
                                                },
                                                "phone": {
                                                    "description": "Номер телефона пользователя",
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "description": "Имя пользователя",
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "email": {
                                                    "description": "Email пользователя",
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "avatar": {
                                                    "description": "Аватар пользователя",
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "description": "Дата и время создания пользователя",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "description": "Дата и время последнего обновления пользователя",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "user_type": {
                                                    "description": "Тип пользователя",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "driver": {
                                            "description": "Данные водителя",
                                            "properties": {
                                                "id": {
                                                    "description": "ID водителя",
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "description": "ID пользователя",
                                                    "type": "integer"
                                                },
                                                "city_id": {
                                                    "description": "ID города",
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "description": "Дата и время создания водителя",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "description": "Дата и время последнего обновления водителя",
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Список ошибок валидации",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Аутентификация пользователя или регистрация нового",
                "description": "Аутентификация пользователя или регистрация нового",
                "operationId": "loginOrRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "description": "Номер телефона пользователя",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "code": {
                                        "description": "Код аутентификации",
                                        "type": "integer",
                                        "example": 1234
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная аутентификация или регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "Токен аутентификации",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Список ошибок валидации",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход пользователя из системы",
                "description": "Выход пользователя из системы",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Пользователь успешно вышел из системы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь успешно вышел из системы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Недопустимый токен аутентификации"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/code": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Создание и отправка проверочного кода на указанный номер телефона",
                "description": "Создание и отправка проверочного кода на указанный номер телефона",
                "operationId": "CreateAndSendCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "description": "Номер телефона пользователя",
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запрос успешно выполнен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Успешность операции",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Список ошибок валидации",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}